Incorporated feedback on #6494
