+++ integration-tests/tests/src/utils/print-warning-box.ts
@@ -1,0 +1,28 @@
+////////////////////////////////////////////////////////////////////////////
+//
+// Copyright 2024 Realm Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+////////////////////////////////////////////////////////////////////////////
+
+export function printWarningBox(...lines: string[]) {
+  const contentWidth = Math.max(...lines.map((line) => line.length));
+  const bar = "━".repeat(contentWidth + 2);
+  console.warn(`┏${bar}┓`);
+  for (const line of lines) {
+    const padding = " ".repeat(contentWidth - line.length);
+    console.warn(`┃ ${line}${padding} ┃`);
+  }
+  console.warn(`┗${bar}┛`);
+}
